requirements.txt
appdirs==1.4.4+computecanada
attrs==23.2.0+computecanada
biopython==1.81+computecanada
certifi==2024.2.2+computecanada
charset_normalizer==3.2.0+computecanada
ConfigArgParse==1.7+computecanada
connection_pool==0.0.3+computecanada
datrie==0.8.2+computecanada
docutils==0.20.1+computecanada
dpath==2.1.6+computecanada
fastjsonschema==2.18.1+computecanada
gitdb==4.0.11+computecanada
GitPython==3.1.40+computecanada
humanfriendly==10.0+computecanada
idna==3.4+computecanada
Jinja2==3.1.3+computecanada
jsonschema==4.20.0+computecanada
jsonschema_specifications==2023.11.2+computecanada
jupyter_core==5.5.0+computecanada
MarkupSafe==2.1.2+computecanada
nbformat==5.9.2+computecanada
numpy==1.24.2+computecanada
packaging==23.2+computecanada
pandas==1.5.3+computecanada
plac==1.3.5+computecanada
platformdirs==3.10.0+computecanada
psutil==5.9.4+computecanada
PuLP==2.7.0+computecanada
python_dateutil==2.9.0.post0+computecanada
pytz==2024.1+computecanada
PyYAML==6.0+computecanada
referencing==0.31.1+computecanada
requests==2.31.0+computecanada
reretry==0.11.8+computecanada
rpds_py==0.8.8+computecanada
smart_open==6.4.0+computecanada
smmap==5.0.1+computecanada
snakemake==7.32.3+computecanada
stopit==1.1.2+computecanada
tabulate==0.9.0+computecanada
throttler==1.2.2+computecanada
toposort==1.10+computecanada
traitlets==5.13.0+computecanada
urllib3==2.1.0+computecanada
wrapt==1.15.0+computecanada
yte==1.5.1+computecanada
zipp==3.17.0+computecanada
